releasing version 0.4
git-svn-id: https://plugins.svn.wordpress.org/websitedefender-wordpress-security/trunk@437308 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ uninstall.php
@@ -1,4 +1,4 @@
-<?php if (!defined('WP_UNINSTALL_PLUGIN')) {exit;} ?>
+<?php if( !defined( 'ABSPATH') && !defined('WP_UNINSTALL_PLUGIN') ) { exit(); } ?>
 <?php
 /*
  * Uninstall plug-in
@@ -7,11 +7,9 @@
  * @since v0.1
  */
 
-/*
- * Delete stored options from the options table
- */
-delete_option('WSD-TARGETID');
-delete_option('WSD-COOKIE');
-delete_option('WSD-TOKEN');
-delete_option('WSD-USER');
-delete_option('wsd_feed_data');
+// this is a shared option so if there are
+// any other plug-ins installed, then don't
+// delete it.
+$__1 = ABSPATH.'wp-content/plugins/secure-wordpress';
+$__2 = ABSPATH.'wp-content/plugins/wp-security-scan';
+if (!is_dir($__1) || !is_dir($__2)) { delete_option('wsd_feed_data'); }
